+++ noodles-sam/src/record/quality_scores.rs
@@ -63,7 +63,8 @@ impl<'a> crate::alignment::record::QualityScores for QualityScores<'a> {
         const OFFSET: u8 = b'!';
 
         Box::new(self.as_ref().iter().map(|&b| {
-            Ok(b - OFFSET) // TODO: Check underflow.
+            b.checked_sub(OFFSET)
+                .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidData, "invalid score"))
         }))
     }
 }
@@ -71,5 +72,31 @@ impl<'a> crate::alignment::record::QualityScores for QualityScores<'a> {
 impl<'a> AsRef<[u8]> for QualityScores<'a> {
     fn as_ref(&self) -> &[u8] {
         self.0
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::alignment::record::QualityScores as _;
+
+    use super::*;
+
+    #[test]
+    fn test_iter() -> io::Result<()> {
+        let quality_scores = QualityScores::new(b"NDLS");
+
+        assert_eq!(
+            quality_scores.iter().collect::<io::Result<Vec<_>>>()?,
+            [45, 35, 43, 50]
+        );
+
+        let quality_scores = QualityScores::new(&[0x00]);
+
+        assert!(matches!(
+            quality_scores.iter().collect::<io::Result<Vec<_>>>(),
+            Err(e) if e.kind() == io::ErrorKind::InvalidData
+        ));
+
+        Ok(())
     }
 }